ci: vitest workspaces (#868)
## 🧰 Changes

Messing around with [Vitest
Workspaces](https://vitest.dev/guide/workspace.html) and it seems pretty
great. Previously, we had to `cd` into the correct subpackage in order
to run `npx vitest [path to file]` but now we can just do it from the
root 😌

I forked a test in `oas-normalize` into its own config + test file and
made it so it uses `chdir` as part of the test. That way, we can run
tests from the both the root and from the subpackages and it'll ✨ just
work ✨

<details><summary>(no longer applicable)</summary>

the downside here is that in `oas-normalize` there's one test that will
now fail unless it's being run from the project root. Feels like a
worthy trade-off, but let me know if you disagree.

</details>

As part of this, I also upgraded our Vitest deps to the latest and added
coverage into the mix.

## 🧬 QA & Testing

If tests (which should be faster now?) pass, we should be in good shape!