fix(exampleGroups): add code sample index (#874)
## 🧰 Changes

We augment the originally defined `x-readme.code-samples` array quite a
bit to get it into the `ExampleGroups` map — for performance (and code
clarity) reasons I think it makes sense to also keep track of where the
code sample was defined in the original array.

## 🧬 QA & Testing

Are we cool with this? Do the updates + snapshot updates look solid?