refactor: remove auto generated min/max values (#881)
| 🚥 Resolves ISSUE_ID |
| :------------------- |

## 🧰 Changes
This removes the auto generated minimum and maximum values we have for
recognized number formats. We're going to be adding labels
[here](https://github.com/readmeio/readme/pull/12148) and it looks
pretty bad for the formats with [longer min max
values](https://github.com/readmeio/readme/pull/12148#issuecomment-2181535819).

We're going to remove this and add this work into the form. 

## 🧬 QA & Testing

Provide as much information as you can on how to test what you've done.