fix(oas): don't add defaults to enum param descriptions (#889)
| 🚥 Resolves CX-947 |
| :------------------- |

## 🧰 Changes

With our new "Defaults to" copy that we're surfacing for request
parameters, we now have duplicated info for parameters that are both
enums and contain a default value ([bin site preview is
here](https://bin.readme.com/s/669ea01aef9101006ed22937)):

![CleanShot 2024-07-22 at 13 16
33@2x](https://github.com/user-attachments/assets/36056d9a-abbe-4283-a15a-a577b8d521d7)


This PR removes that description-updating logic from our
`openapi-to-json-schema` handling.

## 🧬 QA & Testing

Is this the right spot to make this change? This logic is only for
request parameters right?