feat: add Operation.isWebhook() (#898)
| 🚥 Resolves N/A |
| :------------------- |

## 🧰 Changes
I personally find myself checking if the current Operation is a Webhook
or not. I looked at the `Operation` class and didn't see anything like
this. Thought it'd be useful so we don't have to keep type checking in
the main repo.

I put the check on the Operation class since a Webhook will always be an
Operation class but an Operation class might not be a Webhook or
Callback class.

## 🧬 QA & Testing

Provide as much information as you can on how to test what you've done.

---------

Co-authored-by: Jon Ursenbach <erunion@users.noreply.github.com>