feat(oas): add ref name to analyzer (#918)
| 🚥 Resolves RM-11499 |
| :------------------- |

## 🧰 Changes

Add a check to the analyzer if the definition previously had a reference
in it. Once we dereference a file we add `x-readme-ref-name` as a way to
keep track of the previous name. This surfaces if that extension exists
in the definition being analyzed.

## 🧬 QA & Testing

Tests should pass, and analyzer should have a new `x-readme-ref-name`
field being returned.

---------

Co-authored-by: Jon Ursenbach <erunion@users.noreply.github.com>