feat(oas-normalize): reworking how `.validate()` works, new `.convert()` method (#920)
## 🐳  Context

Within our new Git backend we've run into a fun problem where with a
refactor I'm working on to stop caching upconverted and dereferenced
OpenAPI definitions in a database cache, instead opting to pull them
straight from our Git filesystem, I have unfortunately did a number on
an internal endpoint of ours' performance by reducing it from 300+
ops/sec to ~30.

The reason this happened is because in order for this endpoint to
continue to return upconverted OpenAPI definitions (dereferencing is no
longer wanted there) I have to run the file we have in Git, which may be
an OpenAPI, Swagger, Postman, or YAML file, through the `.validate()`
method in `oas-normalize` in order to do this conversion.

Because I already know that these API definitions are valid I just want
to convert, however unfortunately `oas-normalize` doesn't have any
available methods for this.

## 🧰 Changes

> [!IMPORTANT]  
> These are breaking changes to `oas-normalize`.

In order to resolve the above conundrum I am heavily reworking how the
existing `.validate()` method works to now **only** validate a given API
definition. It will continue to convert Postman collections to OpenAPI
in order to validate them as OpenAPI definitons (because we don't have a
Postman validator), but will now just return a boolean in the event of
the definition being valid, or throwing an error if it isn't.

For the conversion side of this that `.validate()` no longer handles I
am introducing a new `.convert()` method that will do as it says:
convert a given API definition to an OpenAPI JSON object. If it's
already OpenAPI, and JSON, it'll return what's already there, but if
it's Swagger or Postman it'll do the conversion to OpenAPI there and
return the result of that -- the same behavior that `.validate()`
previously contained.

Additionally with this, the `.validate()` method had a `convertToLatest`
flag that allowed you, if you so wished, to **not** convert a Swagger
definition to OpenAPI after validating it. I am removing this flag
because as far as I could tell we were always supplying
`convertToLatest` (it was the default behavior). Instead in its place,
if you just want to validate a Swagger definition now, and not convert
it to OpenAPI, don't feed it through `.convert()` after running
`.validate()`.

To be honest I always really disliked that `.validate()` would validate
_and_ convert; seemed like a weird dark pattern and potential footgun.

## ⏱️ Benchmarks

Because the work that spawned this was discovered through the process of
writing a benchmark I'm also creating some new benchmarks here for the
`.validate()` and `.convert()` methods. Diffing the current
`.validate()` benchmark versus the new `.convert()` one you can see
pretty clearly that this new `.convert()` method will greatly improve
our performance over the current implementation and usage of
`.validate()`.

#### `.validate()` currently in `main`


![validate-before](https://github.com/user-attachments/assets/400cc2fa-09a8-4e0f-91e0-dfed5799bf0d)

#### `.validate()` now


![validate-after](https://github.com/user-attachments/assets/eb156d80-962d-44dd-9ea4-6f9ecc3b3f84)

#### `.convert()`

![Screenshot 2024-12-19 at 4 11
12 PM](https://github.com/user-attachments/assets/ba45fc3a-845b-46d2-b171-e894d0eadaa9)

---------

Co-authored-by: Kanad Gupta <git@kanad.dev>