feat(oas-normalize): adding support for parser options on `.deref()` (#955)
## 🧰 Changes

I'm running into a bit of a problem when upgrading
[api](https://npm.im/api) to the latest release of `oas` and
`oas-normalize` in because `OASNormalize.validate()` now only returns a
boolean I was previously expecting a fully dereferenced API definition
there. Unfortunately in changing `.validate()` to return only a boolean
we lost the ability to dereference at the same time, as well as
supplying parser options for dereferencing in order to ignore circular
`$ref` pointers. This adds support for that back in.

* [x] Adding support for supplying `ParserOptions` to a new `parser`
option on the main `OASNormalize` class.
* [x] Deprecating `.deref()` in favor of `.dereference()`
* [x] Deprecating the `parser` option on `.validate()` in favor of the
new `parser` option on the class.
* [x] Updating `.bundle()` to also take advantage of this new `parser`
option to supply options down to that as well.