feat(har-examples): dual ESM and CJS build support (#960)
## 🧰 Changes

I would like to enable `strict` and `noImplicitAny` across this
repository but this `har-examples` package of ours has no typings so it
always gets picked up by TS as not being able to be imported unless we
do a module declaration.

This resolves all of that by migrating the library to full TS, with
actual Har typing enforcement, and a dual CJS and ESM build process.