chore(oas-to-snippet): mock out `httpsnippet-client-api` behaviors (#961)
## 🧰 Changes

Both `oas-to-snippet` and
[httpsnippet-client-api](https://npm.im/httpsnippet-client-api)
depending on `oas` has always made it difficult to do breaking releases
of `oas` because after publishing `oas`, which automatically bumps
`oas-to-snippet`, I then need to bump `httpsnippet-client-api` and
_then_ pull that back into `oas-to-snippet`.

Nuts to all that. Since we are only loading it in tests to test out
behavior for [api](https://npm.im/api) code snippets we can instead just
mock out plugin examples and test around that.