feat(oas): add support for displaying definition size in analyzer (#982)
| 🚥 Resolves ISSUE_ID |
| :------------------- |

## 🧰 Changes

This pull request enhances the OpenAPI analyzer by adding functionality
to calculate and report the raw and dereferenced file sizes of an API
definition.

### Enhancements to Analyzer Functionality:
* Added support for calculating the size of the raw OpenAPI file
(`rawFileSize`) and the dereferenced OpenAPI file
(`dereferencedFileSize`) in MB. These values are now included in the
general analysis results.

### Documentation Updates:
* Updated the `README.md` to include descriptions of the new
`rawFileSize` and `dereferencedFileSize` metrics in the analyzer output.

### Test Coverage:
* Added snapshot tests to validate the inclusion of `rawFileSize` and
`dereferencedFileSize` in the analyzer output.
* Introduced unit tests for `fileSize` and `fileSizeDereferenced` to
verify their correctness.

---------

Co-authored-by: Jon Ursenbach <jon@ursenba.ch>