chore(deps): upgrading `@apidevtools/json-scheam-ref-parser` (#988)
## 🧰 Changes

This upgrade `@apidevtools/json-schema-ref-parser` to the latest release
and also adds a test for the problem that was surfaced in
https://github.com/readmeio/rdme/issues/1306 and fixed in
https://github.com/APIDevTools/json-schema-ref-parser/pull/393.

Since this upgrade had to be reverted earlier today due to it being
released in a patch release, which was automatically pulled down to all
existing rdme versions, since this is a major version bump of
`json-schema-ref-parser` I'll do a major version bump of our parser
library too. Better safe than sorry.