feat(oas): exposing utilities for operationId management (#990)
## 🧰 Changes

This exposes our `getOperationId` and `hasOperationID` methods that are
a core part of `oas` `Operation` class to be usable outside of the
class. I'm doing this because I would like to be able to invoke
`getOperationId` without the memory overhead of instantiating new
instances of `Oas` and `Operation` to do so.