feat(oas-normalize): new option to return validation errors instead of throwing (#1018)
## 🧰 Changes

This expands the `.validate()` method within `oas-normalize` to create a
new `throwIfInvalid` option that will allow us to always return a
`ValidationResult` back for invalid API definitions, instead of always
throwing a `ValidationError`. The reason I am adding this is because I
need to be able to separate errors from warnings in this error as the
current `ValidationError` message is being run through the
`compileErrors` utility from `@readme/openapi-parser` that transforms
errors and warnings into a single human-readable string.

To complement this new `throwIfInvalid` option I am also extending the
existing `ValidationError` exception that's thrown to now contain an
`errors` and `warnings` property for folks who do want exceptions thrown
but still want separate access to this data.

---------

Co-authored-by: Kanad Gupta <git@kanad.dev>
Co-authored-by: Julia Hotaling <42787244+julshotal@users.noreply.github.com>